Move __const.py to utils package
